implement idling for NodePort services
apply the same logic for NodePort that is being used for ClusterIP
services.

If the service can be idled, i.e. has the annotations and idling
is globally enabled in OVN:
if it has endpoints, delete the vip from the idling loadbalancer
and create/update it in the normal loadbalancer, and viceversa if
it has no endpoints.

Signed-off-by: Antonio Ojea <aojea@redhat.com>