Add constants file for e2e package
using a const from a file appended with _test.go will not be
seen at compile time. It will also show up as an error in
an IDE. It was only affecting podNetworkAnnotation in util.go
at this point, but going forward we can centralize all e2e
constants this way.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>