Intoduce ovnkube Node Mode Config
In a Smart-NIC environment, ovnkube node
is required to perfrom different operations
depending on whether it runs on the host or on
the embedded CPU cores of the Smart-NIC.

This commit introduces a new configuration knob
for ovnkube node identified by `ovnkube-node-mode`
flag.

ovnkube-node-mode will specify the mode of operation
which ovnkube-node will operate.

We introduce three modes of operation:

mode="full"
  No modification to ovnkube Node logic

mode=smart-nic-host:
  ovnkube node runs on smart-NIC host, in this mode
  essentialy just setup logging and create and run
  cniserver in smart-nic-host mode.

mode=smart-nic:
  ovnkube node run on Smart-NIC embdedded CPU.
  In this mode Perform the same initialization during ovnkube Node Start()
  however, do not start cni server and instead watch for
  pod events to peform the neccessary plumbing on the smart-nic
  according to special smart-nic anntoation on pod.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>