Create a podman image to use with CI so linter/utests can be run
For OpenShift CI, we need to create a custom image that has podman
installed in it to run the containerized linter, gofmt, and utests
There is no base image that will provide podman by default. So the
alternative is to create our own image and use it in the tests.

Signed-off-by: Aniket Bhat <anbhat@redhat.com>