nics to bridge should take all the IP Families
OVNKube node takes over the default interface and moves
its IPs and Routes to the OVS bridge.

It can happen that the host has IPv6 globally disabled, but
the default interface has it enabled. In this case, we need
to enable IPv6 on the OVS bridge to be able to move the interface
IPv6 addresses and routes.

It is possible that the system, specially with  IPv6, use multipath
routes or a GW that is a non global unicast address.

Signed-off-by: Antonio Ojea <aojea@redhat.com>