deflake e2e test for egressip
the main source of flakiness on e2e test are caused by asynchronous
operations.
Adding fixed timeouts is useless, you'll never now how long can
it takes, and if you overstimate you waste a precious time to run
the test, and also hides race problems in the code.

kubernetes has the wait.PollImmediate functions to check the result
of this asynchronous operations, they give an error if the condition
was not satisfied during the timeout configured.

Signed-off-by: Antonio Ojea <aojea@redhat.com>