Fixes dual-stack reject ACL with empty-lb-events
empty-lb-events adds an lflow in OVN that takes precedence over reject
ACLs. Therefore if an LB in OVN has a VIP for a service on it, the
packet will always generate the empty-lb-event and never reach the
reject ACL. The legacy service and endpoints controller works because on
service add with no endpoints, we dont add the VIP to the LB. Dualstack,
however was always adding VIP to LB.

This patch changes the dualstack behavior to not add VIP on service
create with no endpoints (same as single stack). In the future reject
ACLs will be removed in favor of LB reject, so this is a temporary fix.

Signed-off-by: Tim Rozet <trozet@redhat.com>