retry on unit tests failures
unit test has been flaky for a long time and there is
no sympton that they will be fixed soon.

We can implement the same ginkgo mechanism to retry a few times
in case of failure, to deal with the flaky tests.

This is not great, but we were doing it manually, retesting the PRs,
so at least the developer experience is improved and the test
failures will be more evident.

Signed-off-by: Antonio Ojea <aojea@redhat.com>