De-flake unit tests
The hybrid overlay test: sets up and cleans up a Linux node with a OVN hostsubnet annotation
was flaking because those tests don't use NewLooseCompareFakeExec and it didn't wait for the
add commands to be evaluated before proceeding to the delete part of the test. This patch
does that.

The egress IP test: `should re-assign EgressIPs and perform proper OVN transactions when pod
is created after node egress label switch` was flaking because the retrieved status might be
empty while the tests are running, in which case we had panics observed during the test
execution. Thus check the status length

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>