node deletion tasks on ovnkube manager node is best effort
The clean up of OVN Logical Topology and other runtime state that we do
on ovnkube is best effort. in deleteNode() callback we were doing both,
-- continuing on some errors
-- returning on some errors
As it stands today, the cleanup is best effort as it stands today. So,
the code needs to be consistent in what it is doing.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>