don't compare endpoints on nodes
The nodes need to watch the endpoints in order to delete the stale
conntrack entries.

We were using DeepEqual to compare the endpoints structures, however,
this adds a big penalty that is unneccesary since we are going
to compare both the new and old Endpoints object later to obtain
only the new entries.

Signed-off-by: Antonio Ojea <aojea@redhat.com>