Resync node management port on subnet change
Today when we process a node update we don't check if there has
been a change to the node subnet when determining if we should
resync the management port config in the OVN DB. This will lead
to stale data in the OVN DB and networking not fully working on the
affected node(s). This patch does that.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>