Move lb_force_snat_ip to use router_ip
Previously we were always SNATing to the join facing subnet IP of the
gateway router when any traffic hit a load balancer. With the new
host->service implementation using openflow, the traffic in certain
cases will be hairpinned back out of the GR towards br-ex. In this case
we need the packet to be SNAT'ed to the GR external facing IP address
and not the internal join subnet IP.

This patch leverages new OVN behavior where by specifying "router_ip",
OVN will automatically SNAT to the egress interface IP as the packet
leaves the GR.

Signed-off-by: Tim Rozet <trozet@redhat.com>