External gateway namespace tests: Cover BFD
We clone the external gateway namespace annotation tests with a difference:
bfd is enabled both on ovn side and on the containers
(using a custom container via frr).

We check that ecmp works with two containers, then we remove one
container and check that all the traffic reaches the remaining
container.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>