Gateway tests: collect all the goroutines before exiting
The ping command on pods may be slower that the detection of the icmp
message on the targets. In this case the goroutines are still running
but the target is brought down, causing the test to fail. Here we add a
waitgroup to ensure all the pingers are executed before turning down the
target.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>