Fixes mismatch in DGP MAC for ipv6 deployments
With 380e80f389dad997f9f97c030eb89d8abce4204f we changed the MAC for DGP
on the OVN side to always be the MAC determined by IPv4 address.
However, the corresponding change was not made for the OVS interface.
Thus IPv6 DGP would send packets to a MAC that did not exist on the
system and traffic would be dropped.

Signed-off-by: Tim Rozet <trozet@redhat.com>