Fix: egress IP flakes, wrap update in mutex execution
Egress IP is still flaking in CI. I believe the cause of this is due
to not wrapping updating the egress IP status in the retry assignment
executin. On slow servers there can be a conflicting update when the
nodes are labelled and the egress IP is added during our e2e tests.
Wrapping the update part in the mutex execution will enforce that only
one update is done at a time, either when a node is added and the egress
IP is re-assigned, or when the egress IP is added / updated.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>