Fixes errors when adding NAT with existing entry
When adding SNAT per pod with an existing entry, the code will fail and
report an error because with OVN we cannot do lr-nat-del and lr-nat-add
in the same OVSDB transaction. This is because OVSDB evaluates the
lr-nat-add and thinks there will be a conflict because an entry already
exists.

Additionally, during gateway init for a node, this failure may also
occur. In which case the GR will be left with missing config.

To handle this case, split add and delete into separate transactions,
and only del/add if the entry is missing.

Signed-off-by: Tim Rozet <trozet@redhat.com>