sflow configuration's agent column should use default setting
the code currently sets the agent IP to ovn-k8s-mp0, which makes all
the sFlow traffic to go over the OVN logical network and it's incorrect
(unless the collector itself is running as a Pod).

the agent column is optional and with the default setting ovs-vswitchd
daemon running on the K8s Host/Node is going to send the sFlow sampled
packets out through the interface as provided by the routing table based
on the first collector IP.

if the collector is a Pod, then it will choose ovn-k8s-mp0. on the other
hand if the collector is on th eunderlay, then it will choose K8s node
IP

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>