Unidling: make the wait shorter / check for longer time
When the service times out but it should not, 5 seconds are not enough
for the eventually. 3 seconds are consumed by the first timeout, and it
could exceptionally happen that the second attempt lasts longer than 2.
Here we lower the timeout and wait for longer for the right config to
get in place.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>