Add routes for other host IP addresses
For shared gateway mode, pods cannot get to other IPs that are on the
host without logical_router_policies. Today we can get to the node via
the kube nodeIP, but not other IPs on the node. This fixes it by reading
the host ip addresses on the node annotation, and adds routes.

Signed-off-by: Tim Rozet <trozet@redhat.com>