Idling test: Wait one second before generating a new event
The way we check if hitting an idling service generates an idling event is
by checking the last event time, hitting the service and checking if an
event is created with a bigger time. If this happens within the second,
the two events will have the same time and the test will flake.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>