pods: add short node name hash to pod logical switch ports
If ovn-kube node on a given node is wedged or crashlooping and a
pod is deleted and rescheduled on another node, the OVS interface will
still have the pod's iface-id. But now the new node will also have
the new pod's iface-id, and the ovn-controllers will fight over
the port binding of the pod.

Adding the node name to the logical switch port name ensures that no
two controllers will ever fight over the same pod's LSP.

TODO: figure out the backwards compat story

Signed-off-by: Dan Williams <dcbw@redhat.com>