CNI cmdCheck: treat ingress_policing_rate=0 as not found
When the cni is not chained with multus, the container runtime invokes
the check command. The check command checks the bandwith requirements
annotated on the pod, vs what's being configured in ovs. Ovs sets
ingress_policing_rate=0 by default, so the check always fails.

Here, when ingress_policing_rate is 0, we treat it as not found.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>