Purge only the lprs with nexthops of the same family of new nexthop
For each node, we add pbrs for each node subnet. When syncing pbrs, we purge
not expected rules. In case of dual stack, we are mistakenly purging the
rules that were added because of the subnet belonging to the other ip
family.
Here, we compare the family of the nexthop with the family of the new
nexthop and apply the logic only if they match.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>