sriov: support virtio vdpa devices
Add support to sriov-based virtio-vdpa devices.

When a sr-iov device has a vdpa device created, and it is bound to
virtio_vdpa driver, use the associated virtio-net device instead of the
vendor one.

In order to determine whether a device is vdpa or not as well as its
device path, use external
library github.com/k8snetworkplumbingwg/govdpa

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>