Implement ETP: Per node br-ex flow update + IPtables changes
UPSTREAM COMMIT: bd7a7f6e8ee140a9988c81e7fa438e95826f39e7

This Commit focuses on the case for services with
ExternalTrafficPolicy=Local where traffic is hitting a svc with
host networked backends from outside the cluster and from the
host itself. It has a few major functions

- Adds an endpoint watcher to the `gateway_shared_intf.go` code
to ensure we correctly manage any "special flows" for the case
stated above

- It adds 4 new flows to br-ex in ct Zone 64003 for each nodeport,
ExternalIP or Loadbalancer type service that match the above
criteria. These new flows allow the traffic to flow outside of
OVN-K's control.

1. A Flow to DNAT traffic that arrives at eth0 with a destination
(Port or IP) of a relevant service to the correct host networked
backend (Loadbalancing is not guarenteed with ETP=Local. The flow
concludes by pushing to 2.

2. A Flow to ouput the DNATed packet to the LOCAL netwoking stack
since dst_ip is now that of node

3. A flow to match on return traffic coming from the Host Networked
backend pod based on the relvant port. It proceedes to Un-DNAT the
traffic back to the correct IP.  Then it pushes the packet to 4.

4. A flow to take the reply traffic and send it directly out eth0 of
the node back to it's src

- Manupulates IPtables rules for the referenced case so that the
host -> service -> host-networked-pod flow works correctly. Here since
Loadbalancing is not guarenteed with ETP:Local we can just ensure
the incoming packet is DNATed to the correct nodeIP:Port

Signed-off-by: astoycos <astoycos@redhat.com>