Reduces number of remove VIP operations in svcs
We see at scale that we are calling remove VIP way too many times due to
the fact that we have to ensure that the VIP does not exist on another
load balancer when we update it. We were blindly doing that for idling
and the cluster vs worker loadbalancers.

This patch addresses it by using a cache inside service tracker to track
which type of load balancer the VIP is currently on. If the entry
doesn't exist, we brute force remove as before, but if it does exist we
try to avoid unnecessary nbctl commands to remove.

Note, using go-ovn or libovsdb bindings would be a better solution here,
but that implementation can come later as it will require a much bigger
change to the project.

Signed-off-by: Tim Rozet <trozet@redhat.com>