Batch service delete OVN transactions
When we have to run delete commands across all nodes, it results in many
nbctl remove vip commands twice for every single worker. This patch
batches the remove commands into a single nbctl command, reducing the
times we have to invoke nbctl.

In the future when we switch to go bindings, this is still useful in
order to batch transactions there as well for a service.

Signed-off-by: Tim Rozet <trozet@redhat.com>