LGW: Lookup node IPs on every NP/LB/EIP service event
This Commit ensures that the correct IPtables rules are added
in the case that a secondary node IP is added after ovnkube-node
has already been started.

It simply looks up the node ips for each service event rather than
storing the ips in memeory.

The Commit also ensures we call the net function `InterfaceAddrs()`
with our mocking struct for easier unit testing

Signed-off-by: astoycos <astoycos@redhat.com>