LGW: Lookup node IPs on every NP/LB/EIP service event
This Commit ensures that the correct IPtables rules are added for
externalIP services in the case that a secondary node IP was added
after ovnkube-node had already been started.

It simply looks up the node ips for each service event rather than
storing the ips in memory.

On local gateway mode, we only install iptables rules for ExternalIPs
when the IP is present in the node.
However, on ovnkube-node restart, we were using the same function
that shared-gateway, that doesn't take into account if the ExternalIP
is present, hence installing iptables rules for those ExternalIPs.

Signed-off-by: astoycos <astoycos@redhat.com>