Batching: Fixes finding maximum bash arguments
It turns out what the system reports for maximum arguments may not be
the real amount of arguments that are usable in a shell. Different
kernels calculate it differently, but the gist is that the only reliable
way to know is to test for it. This patch does an initial search based
on the given max args in the kernel to find the actual usable amount.

The search is only done one time during initialization, so it shouldn't
impact performance.

Signed-off-by: Tim Rozet <trozet@redhat.com>