support for VLAN for the OVS bridge flows that replaced DGP
currently the set of OpenFlow rules that were added to replace DGP
assumes that the VLAN is native and untagged. However, in the case of
BF2 it could be that the BF2 Host could be tagged and BF2 admin VLAN
is native.

(cherry picked from commit 3fe2e5bd8e24c53571dbaf68b168824902e70a03)
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>