Declare a maximum line length for batching
There are still bash errors for "too many args" which stem not from the
number of arguments but a combination of number of args and total length
of the args. This sets the maximum line length to 100k, which is less
than the default maximum of 130k (to be safe).

Signed-off-by: Tim Rozet <trozet@redhat.com>