Remove --disable-snat-mutliple-gw check for host-level OVS flows
This setup makes sure that we route pod related traffic (coming from
ICNI) into OVN. We can however always peform this setup without needing
to check if `--disable-snat-multiple-gw` is enabled. The reason for this
is that the NAT rules on the GW router in OVN specify "snat", and given
the first commit in this patch: will always do so now. If reply ICNI traffic
hits a node with the destination IP being a pod IP: we will steer the
traffic into OVN. If an external client sends a packet to a pod IP we
will drop the packet (which is correct, because no external connections
should be possible directly to a pod IP) because NAT rules specifying
"snat" only accepts incoming packets for return traffic, and drops
everything else.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>