Detect if OVS supports check pkt length
In older kernels like RHEL 7, OVS does not support chk pkt length in
kernel datapath. This will cause all packets hitting these flows to go
to userspace, which is the majority of OVN destined ingress traffic into
the node. This will result in severe performance degradation. To avoid
this, detect and disable using these flows if the kernel datapath does
not support the action.

Note this will result in these nodes not being able to handle packets
larger than pod MTU and send ICMP frag needed, but the trade off vs.
performance impact is worth it.

Signed-off-by: Tim Rozet <trozet@redhat.com>