Route Pod network traffic to host through tunnel
Today pod network traffic to k8s workers(host network)
is sent through GR to be SNATed and then to the underlay.
on the recieving end, traffic is sent to CT on the
physical bridge and forwarded to the host
i.e LOCAL port or host representor port in case of Smart-NICs.

By design, the conntrack state on zone 64000 is never
established for host destined traffic which may prevent
offload capable NICs from offloading this type of traffic.

To overcome the issue, we will route host destined traffic
from OVN to the respective ovn management port.
doing so, we trade the additional conntrack on zone 64000
(on both sender and reciever side) and SNAT of GR (on sender side)
with tunnel encap/decap.

As many NICs are capable of offloading tunnel encap/decap
operations, it will result in an overall increase in performance
where supported hardware is present.

This commit modifies the existing policy based routes on
ovn cluster router and adds a new config knob
"EnableTunneledHostTraffic" to control the routing of host
destined traffic from pod network.

When enabled, Policy based routes are addeded to route host destioned
ip traffic to the respective host through its management port(ovn-k8s-mp0).

When disabled, only pod to node local traffic will be routed through
the management port as it is today.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>