Adds back logical ports for pods during node add
When a node is deleted, there is a grace period before pods will be
evicted from that node. A scenario can occur where we get a node delete
followed by a node add for the same node. During node delete we tear
down the logical switch and all ports associated with that node,
including pod ports. If the above scenario happens, the logical switch
will be restored, but existing pods will not get any new add event, so
we will not recreate any of the logical ports. This will leave the
container without networking and on subsequent restart the pod wil be
stuck in containerCreating status.

This patch addresses the issue by ensuring all pods on a node are wired
after a node add.

Signed-off-by: Tim Rozet <trozet@redhat.com>