Shared gateway use ct_mark to classify host/ovn traffic
Today, openflow rules on the shared gateway bridge
send incoming traffic from uplink to conntrack on zone
64000 to decide if the traffic needs to be steered towards
OVN or the host.

established/related connections are sent to OVN via
the patch port to OVN integration bridge while new
connections are set to the host.

This poses an issue for NICs that are capable of offloading
conneciton tracking rules as only established connections are
offloadable.

This commit modifies the openflow rules on the physical
bridge to allow the use of ct_mark on ct zone 64000 to
classify traffic going to host and OVN

- table 0: set up higher priority rule (205/200) to steer tunneled traffic
  based on udp encap port
- table 0: commit traffic coming out of ovn (patch to br-int) to zone 64000
  with ct_mark=0x1
- table 0: commit traffic coming from host (LOCAL) to zone 64000 with
  ct_mark=0x2
- table 1: send established/related connections on zone 64000 to
  either OVN or host based on ct_mark

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>